Include styles to be consistent with the custom component refactor
