+++ src/components/Dashboard.module.scss
@@ -8,14 +8,7 @@
       margin: 0 auto;
       padding: 1rem;
     }
-  }
-  
-  .empty {
-    display:flex;
-    align-items: center;
-    
-  }
-  
+  } 
   .app__brand {
     color: #3cd5ff;
     font-style: italic;
@@ -30,65 +23,4 @@
     padding: 1rem;
   }
   
-  .widget {
-    display: flex;
-    flex-direction: column;
-    color: white;
-    background-color: #1a2233;
-    filter: drop-shadow(10px 10px 20px rgb(0 0 0 / 25%));
   
-    &__header {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      padding: 0.5rem;
-      background-color: #3cd5ff;
-    }
-  
-    &__status {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-    }
-  
-    &__title {
-      overflow: hidden;
-      color: #1a2233;
-      font-weight: bold;
-      font-size: 0.875rem;
-      white-space: nowrap;
-      text-decoration: none;
-      text-overflow: ellipsis;
-    }
-  
-    &__body {
-      flex: 1;
-      padding: 0.5rem;
-    }
-  
-    &__description {
-      overflow: hidden;
-      text-overflow: ellipsis;
-      display: -webkit-box;
-      -webkit-line-clamp: 2;
-      line-clamp: 2;
-      -webkit-box-orient: vertical;
-    }
-  
-    &__footer {
-      display: flex;
-      align-items: center;
-      justify-content: space-around;
-      padding: 0.5rem;
-    }
-  
-    &__stat {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-  
-      span {
-        margin-left: 0.25rem;
-      }
-    }
-  }
