+++ src/components/ErrorWidget.module.scss
@@ -1,0 +1,7 @@
+  
+  .empty {
+    display:flex;
+    align-items: center;
+    flex-direction: column;
+    color: red;
+  }
